Move stuff around to improve readability.
Focus on requirements and design elements that need to be highlighted.

Update the detailed API flow for accuracy and add more details.

Update security considerations to focus more on the Relay Server specific topics.

separate out security considereations for clients under a sub-section

Mention overall system security considerations and point out that they are out of scope of this document.